Fix menu weirdness on some devices.
Setting the orientation of the game to be landscape seems to cause some
odd behaviour on e.g. a Moto G5S Plus. GLActivity gets launched in
portrait, initialises GBCC, but not the window/menu. Then, the activity
stops, saving its state as if you'd pressed home, rotates, then
restarts properly, loading the just-saved state. As the menu was never
initialised, we now have some uninitialised menu settings being used.
This causes the menu to not initialise properly.

This fix is to store whether the menu was initialised when we save the
state, then check this upon loading.